requirements.txt
azure-functions>=1.21.3
azure-storage-queue>=12.12.0
azure-storage-blob>=12.24.1
